+++ app/controllers/users/sessions_controller.rb
@@ -1,7 +1,17 @@
 module Users
   class SessionsController < Devise::SessionsController
-    skip_before_action :save_current_path
+    include CookiesHelper
+
     skip_before_action :set_current_user
+    skip_before_action :save_current_path
+    before_action :forget_user, only: :destroy
+
+    private
+
+    def forget_user
+      Current.user = nil
+      forget(current_user) if current_user.present?
+    end
 
     protected