+++ db/migrate/20191108073028_add_email_to_identities.rb
@@ -1,0 +1,6 @@
+class AddEmailToIdentities < ActiveRecord::Migration[5.2]
+  def change
+    add_column :identities, :email, :string
+    add_index :identities, [:uid, :provider]
+  end
+end