+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_11_07_122736) do
+ActiveRecord::Schema.define(version: 2019_11_08_083442) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_trgm"
@@ -173,6 +173,17 @@ ActiveRecord::Schema.define(version: 2019_11_07_122736) do
     t.index ["world_id"], name: "index_guilds_on_world_id"
   end
 
+  create_table "identities", force: :cascade do |t|
+    t.integer "user_id"
+    t.string "provider", null: false
+    t.string "uid", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.string "email"
+    t.index ["uid", "provider"], name: "index_identities_on_uid_and_provider"
+    t.index ["user_id"], name: "index_identities_on_user_id"
+  end
+
   create_table "notifications", force: :cascade do |t|
     t.jsonb "name", default: {"en"=>"", "ru"=>""}, null: false
     t.string "event"
@@ -266,6 +277,7 @@ ActiveRecord::Schema.define(version: 2019_11_07_122736) do
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.string "role", default: "user", null: false
+    t.string "remember_digest"
     t.index ["email"], name: "index_users_on_email", unique: true
   end