+++ spec/factories/identities.rb
@@ -1,0 +1,8 @@
+FactoryBot.define do
+  factory :identity do
+    uid { 123 }
+    provider { 'discord' }
+    sequence(:email) { |i| "identity#{i}@gmail.com" }
+    association :user
+  end
+end