+++ spec/models/user_spec.rb
@@ -1,5 +1,7 @@
 RSpec.describe User, type: :model do
   it { should have_many(:characters).dependent(:destroy) }
+  it { should have_many(:guilds).through(:characters) }
+  it { should have_many(:identities).dependent(:destroy) }
   it { should validate_presence_of :email }
   it { should validate_presence_of :password }
   it { should validate_presence_of :role }
@@ -88,6 +90,45 @@ RSpec.describe User, type: :model do
 
         expect(result.size).to eq 1
         expect(result[0].id).to eq static1.id
+      end
+    end
+
+    context '.remember' do
+      let!(:user) { create :user }
+
+      it 'updates remember_digest param from nil' do
+        expect { user.remember }.to change(user, :remember_digest).from(nil)
+      end
+    end
+
+    context '.forget' do
+      let!(:user) { create :user }
+      before { user.remember }
+
+      it 'updates remember_digest param to nil' do
+        expect { user.forget }.to change(user, :remember_digest).to(nil)
+      end
+    end
+
+    context '.authenticated?' do
+      let!(:user) { create :user, remember_digest: nil }
+
+      it 'returns false for nil remember_digest' do
+        expect(user.authenticated?('')).to eq false
+      end
+
+      it 'returns false for wrong digest/token' do
+        token = SecureRandom.urlsafe_base64
+        user.update_attribute(:remember_digest, User.digest(token))
+
+        expect(user.authenticated?(token + '1')).to eq false
+      end
+
+      it 'returns true for correct digest' do
+        token = SecureRandom.urlsafe_base64
+        user.update_attribute(:remember_digest, User.digest(token))
+
+        expect(user.authenticated?(token)).to eq true
       end
     end
   end