+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- omniauth support
+- login through discord
+- save user in session for discord auth
+
 ## [0.6.7] - 2019-11-08
 ### Fixed
 - bug with character updating and lefting guild